Backport sdf6: Move recursiveSameTypeUniqueNames from ign.cc to parser.cc
Backport of sdf9 PR 606 https://osrf-migration.github.io/sdformat-gh-pages/#!/osrf/sdformat/pull-requests/606
Commit: https://github.com/osrf/sdformat/commit/2d1bc5627542f4072b1c2c5628074e8e1120a93d
